Moves snap data to more logical places
