+++ src/data/state/massachusetts/name-cores.js
@@ -9,19 +9,18 @@ const UNEARNED_INCOME_SOURCES = [
   'WorkersComp', 'Pension', 'SocialSecurity', 'Alimony', 'OtherIncome'
 ];
 
-// // More needs to be confirmed about this one
-// const NET_INCOME_DEDUCTIONS = [
-// 	'Standard', 'EarnedIncome', 'Medical', 'DependentCare',
-// 	'AllowableChildSupport', 'AllowableHomelessExpenses', 'ExcessShelterExpense'
-// ];
-
 const CHILD_CARE_EXPENSES = [
 	'ChildDirectCareCosts', 'ChildBeforeAndAfterSchoolCareCosts',
 	'ChildTransportationCosts', 'ChildOtherCareCosts'
 ];
 
+const ADULT_CARE_EXPENSES = [
+	'AdultDirectCareCosts', 'AdultTransportationCosts',
+	'AdultOtherCareCosts'
+];
+
 export {
 	UNEARNED_INCOME_SOURCES,
-	// NET_INCOME_DEDUCTIONS,
-	CHILD_CARE_EXPENSES
+	CHILD_CARE_EXPENSES,
+	ADULT_CARE_EXPENSES
 }
